Rename init.h to Logger.h
