+++ src/software/ai/hl/stp/play/defense_play.cpp
@@ -11,7 +11,7 @@
 #include "software/ai/hl/stp/tactic/shoot_goal_tactic.h"
 #include "software/ai/hl/stp/tactic/stop_tactic.h"
 #include "software/geom/util.h"
-#include "software/logger/init.h"
+#include "software/logger/Logger.h"
 #include "software/parameter/dynamic_parameters.h"
 #include "software/util/design_patterns/generic_factory.h"
 #include "software/world/game_state.h"
