+++ src/software/ai/hl/stp/tactic/defense_shadow_enemy_tactic.cpp
@@ -5,7 +5,7 @@
 #include "software/ai/hl/stp/action/move_action.h"
 #include "software/ai/hl/stp/action/stop_action.h"
 #include "software/ai/hl/stp/tactic/mutable_tactic_visitor.h"
-#include "software/logger/init.h"
+#include "software/logger/Logger.h"
 #include "software/parameter/dynamic_parameters.h"
 
 DefenseShadowEnemyTactic::DefenseShadowEnemyTactic(const Field &field,
