+++ src/software/backend/simulation/simulator_robot_singleton.cpp
@@ -1,6 +1,6 @@
 #include "software/backend/simulation/simulator_robot_singleton.h"
 
-#include "software/logger/init.h"
+#include "software/logger/Logger.h"
 
 extern "C"
 {
