+++ src/software/full_system_main.cpp
@@ -7,7 +7,7 @@
 #include "software/ai/hl/stp/play_info.h"
 #include "software/backend/backend.h"
 #include "software/constants.h"
-#include "software/logger/init.h"
+#include "software/logger/Logger.h"
 #include "software/util/design_patterns/generic_factory.h"
 #include "software/visualizer/visualizer_wrapper.h"
