+++ src/software/logger/BUILD
@@ -3,8 +3,8 @@ package(default_visibility = ["//visibility:public"])
 cc_library(
     name = "logger",
     hdrs = [
+        "Logger.h",
         "custom_logging_levels.h",
-        "init.h",
     ],
     deps = [],
 )
